% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abies}
\alias{abies}
\title{Global distribution of Abies}
\format{A data.frame with 4 columns
'abies'}
\usage{
data(abies)
}
\description{
A sample of GBIF point locality occurrence data for the genus Abies for illustrating findlocal and geo_findlocal routines
}
\keyword{datasests}
