% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cracle_build.R
\name{addplot}
\alias{addplot}
\title{Adds a single PDF plot to already open plot}
\usage{
addplot(dens.ob, var, col = sample(grDevices::colours()), type = ".kde")
}
\arguments{
\item{dens.ob}{An object derived from the vegdistmod::densform() or vegdistmod::and_fun() or vegdistmod::or_fun() functions.}

\item{var}{A character string that matches one of the layer names in the source raster object.}

\item{col}{A color declaration. Default is a random color.}

\item{type}{A character string of value either ".kde" for a Kernel Density Estimator curve, or ".gauss" for a Gaussian (normal) curve. All other values will result in errors.}
}
\description{
Using an object from the vegdistmod::densform() or vegdistmod::and_fun() or vegdistmod::or_fun() functions, add a PDF over the existing plot. Useful for visualizing joint likelihood curves.
}
\examples{
#distr <- read.table('test_mat.txt', head=T, sep ="\\t");
#OR:
data(distr); data(climondbioclim);
#extr.raw = extraction(data=distr, clim= climondbioclim, schema='raw');
#dens.list.raw <- dens_obj(extr.raw, clim = climondbioclim, bw = 'nrd0', n = 1024);
#multiplot(dens.list.raw, names(climondbioclim[[1]]));
#or <- or_fun(dens.list.raw);
#addplot(or, names(climondbioclim[[1]]), col ='black');
#and <- and_fun(dens.list.raw);
#addplot(and, names(climondbioclim[[1]]), col ='black');
}

