% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cracle_build.R
\name{and_fun}
\alias{and_fun}
\title{P(A | B) = P(A) * P(B)}
\usage{
and_fun(dens.oblist, w = FALSE)
}
\arguments{
\item{dens.oblist}{An object derived from the vegdistmod::dens_ob() function.}

\item{w}{Weight importance of probability functions}
}
\description{
Using an object from the vegdistmod::dens_obj() function. Create a single density object (i.e., like that produced by vegdistmod::densform()) where the probability curves correspond to the probability density function of ALL taxa/species from the original set occurring.
}
\examples{
#distr <- read.table('test_mat.txt', head=T, sep ="\\t");
#OR:
data(distr);
data(climondbioclim);
extr.raw = extraction(data=distr, clim= climondbioclim, schema='raw');
dens.list.raw <- dens_obj(extr.raw, clim = climondbioclim, bw = 'nrd0', n = 1024);
multiplot(dens.list.raw, names(climondbioclim[[1]]));
or <- or_fun(dens.list.raw);
addplot(or, names(climondbioclim[[1]]), col ='black');
and <- and_fun(dens.list.raw);
addplot(and, names(climondbioclim[[1]]), col ='black');
}

