% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climondbioclim}
\alias{climondbioclim}
\title{Raster grid for 19 Bioclimatic Variables from the CLIMOND project}
\format{A raster stack}
\source{
\url{https://www.climond.org}
}
\usage{
data(climondbioclim)
}
\description{
A dataset of climate model output for 19 bioclimatic variables. Original data (AND MORE) can be found here (Account required but free): https://www.climond.org
}
\keyword{datasests}

