% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cracle_build.R
\name{cracboot}
\alias{cracboot}
\title{Vegetation resampling CRACLE bootstrap.}
\usage{
cracboot(dens.oblist, clim, n = 100)
}
\arguments{
\item{dens.oblist}{A density list object derived from the vegdistmod::dens_obj() function.}

\item{clim}{A raster file containing environmental data. MUST be the same one used to create the density list object.}

\item{n}{Number of bootstrap iterations. Default is 100.}
}
\description{
Using an density list object (vegdistmod::dens_obj()) draw incomplete samples of taxa from the list of taxa and run CRACLE. The Gaussian and KDE derived maximum likelihood values are kept and summarized as confidence intervals and bootstrap mean and standard deviation for each variable.
}
\examples{
#distr <- read.table('test_mat.txt', head=T, sep ="\\t");
#OR:
data(distr); data(climondbioclim);
extr.raw = extraction(data=distr, clim= climondbioclim, schema='raw');
dens.list.raw <- dens_obj(extr.raw, clim = climondbioclim, bw = 'nrd0', n = 1024);
and <- and_fun(dens.list.raw);
optima <- get_optim(and);
boot.raw <- cracboot(dens.list.raw, clim = climondbioclim, n = 100); 
#Compare to optima.
write_results(optima, clim = climondbioclim, method = 'conintkde', file = 'raw')
write_results(boot.raw, clim = climondbioclim, method = 'conintkde', file = 'boot.raw')
}

