% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cracle_build.R
\name{densplot}
\alias{densplot}
\title{Plot PDF curve of given type and variable for single density object.}
\usage{
densplot(dens.ob, var, col = sample(grDevices::colours()), type = ".kde")
}
\arguments{
\item{dens.ob}{An object derived from the vegdistmod::dens_ob() function.}

\item{var}{A character string that matches one of the layer names in the source raster object.}

\item{col}{A color declaration. Default is a random color.}

\item{type}{A character string of value either ".kde" for a Kernel Density Estimator curve, or ".gauss" for a Gaussian (normal) curve. All other values will result in errors.}
}
\description{
Using an object from the densform() function. Plot a single PDF curve in a new plot window.
}
\examples{
#distr <- read.table('test_mat.txt', head=T, sep ="\\t");
#OR:
data(distr); data(climondbioclim);
extr.raw = extraction(data=distr, clim= climondbioclim, schema='raw');
extr.sub = subset(extr.raw, extr.raw$tax == extr.raw[5,'tax']);
dens.sub = densform(extr.sub, clim = climondbioclim, bw = 'nrd0', n = 512);
densplot(dens.sub, names(climondbioclim[[1]]));
}

