% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distr}
\alias{distr}
\title{A sample of species data}
\format{A data.frame with 5 columns
'distr'}
\usage{
data(distr)
}
\description{
A sample of GBIF point locality occurrence data for species co-occurring at a vegetation plot in SW Virginia
}
\keyword{datasests}

