% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distrgen}
\alias{distrgen}
\title{A sample of genera data}
\format{A data.frame with 5 columns
'distrgen'}
\usage{
data(distrgen)
}
\description{
A sample of GBIF point locality occurrence data for the genera co-occurring at a vegetation plot in SW Virginia
}
\keyword{datasests}

