% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fun.R
\name{filter}
\alias{filter}
\title{Filter a set of occurrence data based on the multivariate likelihoods.}
\usage{
filter(ext_ob, dens_ob, clim, min = 0, alpha = 0.01, type = ".kde")
}
\arguments{
\item{ext_ob}{A data.frame of climate values (i.e., extracted from the climate raster object). Will be passed to multiv_likelihood().}

\item{dens_ob}{A density object from vegdistmod::densform(), vegdistmod::and_fun(), or vegdistmod::or_fun();}

\item{clim}{A raster object of climate data (matching x)}

\item{min}{The minimum likelihood to be kept. Will override the value of alpha given. Optional.}

\item{alpha}{The value of alpha you would like to use for confidence interval construction. Default to 0.01 for a 99 percen confidence interval.}

\item{type}{Designate either ".gauss" or ".kde".}
}
\description{
Given a table including localities and climate data,and a value of alpha this 
function will return the same table with statistical outliers
removed according to the alpha-confidence interval of the likelihoods.
}
\examples{

data(abies);
ext.abies = extraction(abies, climondbioclim, schema='raw');
dens.abies = densform(ext.abies, climondbioclim);
f <- 
 filter(ext.abies, dens.abies, climondbioclim, alpha = 0.01, type = '.kde')

}

