% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fun.R
\name{findlocal}
\alias{findlocal}
\title{Search for likely occurrences}
\usage{
findlocal(ext_ob, clim, type = ".kde", maxiter = 50, searchrep = 3,
  manip = "condi", alpha = 0.05, factor = 4, n = 1024, w = FALSE)
}
\arguments{
\item{ext_ob}{A data.frame of climate values (i.e., extracted from the climate raster object). Will be passed to multiv_likelihood().}

\item{clim}{A raster object of climate data (matching x)}

\item{type}{Designate either ".gauss" or ".kde".}

\item{maxiter}{Maximum number of search iterations.}

\item{searchrep}{How many times to search for simulated localities 
(per parent occurrence) per iteration. Recommend 1, but feel free to tune this parameter.}

\item{manip}{Specify either 'reg', 'condi', or 'bayes. Default is 'condi'.
See documentation for densform() for descriptions of these.}

\item{alpha}{The value of alpha to be used to calculate the initial confidence interval 
for removing climatic outliers in the sample(s).}

\item{factor}{To be passed to the extraction() function for post search 
thinning of data to limit overfitting. Set to 1 to ignore.}

\item{n}{An integer value of the number of bins to use for Kernel Density Estimation}

\item{w}{To weight the log-likelihoods by coefficient of variation or not.}
}
\description{
This is a Markov Chain like search procedure to construct a set of likely occurrences 
for a given sample of known occurrences. Simulated occurrences are accepted if each one 
is more likely than the parent locality used to find the simulated point AND if the 
multivariate likelihood of the entire sample is improved. The improvement of the entire
sample is calculated after each parent locality is visited once in the search process. 
The process is repeated a given number of times (think Markov Chain again), to hopefully 
reach a near optimal set of simulated occurrences.
}
\examples{
\dontrun{

data(abies);
ext.abies = extraction(abies, climondbioclim, schema='flat', factor =4);
sea <- findlocal(
 ext.abies, climondbioclim, 
 type = '.kde', 
 maxiter = 15, searchrep = 1, 
 manip = 'condi');
 plot_clim(sea[[1]], climondbioclim[[5]])
 }
}

