% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_direct.R
\name{gbif_get}
\alias{gbif_get}
\title{Download distribution data directly from GBIF API}
\usage{
gbif_get(taxon, maxrec = 2e+05)
}
\arguments{
\item{taxon}{A string of the form 'genus species' or 'genus'.}

\item{maxrec}{Maximum number of records to download.}
}
\description{
This function requests data from the GBIF database for a single taxon using the GBIF callback API.
}
\examples{
\dontrun{
abies <- gbif_get('Abies');
}
}
