% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fun.R
\name{geo_findlocal}
\alias{geo_findlocal}
\title{Search for likely simulated localities with geographic subsetting.}
\usage{
geo_findlocal(ext_ob, clim, type, maxiter = 10, bg = 0, searchrep = 1,
  manip = "condi", alpha = 0.05, divisions = 10, factor = 4)
}
\arguments{
\item{ext_ob}{A data.frame of climate values (i.e., extracted from the climate raster object). Will be passed to multiv_likelihood().}

\item{clim}{A raster object of climate data (matching x)}

\item{type}{Designate either ".gauss" or ".kde".}

\item{maxiter}{Maximum number of search iterations.}

\item{bg}{An object of background point climate data matching the 
output of extraction(). Generate random backround points and then use extraction().}

\item{searchrep}{How many times to search for simulated localities 
(per parent occurrence) per iteration. Recommend 1, but feel free to tune this parameter.}

\item{manip}{Specify either 'reg', 'condi', or 'bayes. Default is 'condi'.
See documentation for densform() for descriptions of these.}

\item{alpha}{The value of alpha to be used to calculate the initial confidence interval 
for removing climatic outliers in the sample(s).}

\item{divisions}{How many times should the data be split into quadrants? 
Default is 5 resulting in 20 (5x4quads) geographically oriented samples to be selected.}

\item{factor}{To be passed to the extraction() function for post search 
thinning of data to limit overfitting. Set to 1 to ignore.}
}
\description{
This wraps the search() function within a set of commands that randomly splits the distribution 
in to 4 quadrants 'n' number of times. For widespread, genetically diverse species this has 
the effect of applying more locally relevant probability functions to the data.
}
\examples{
data(abies);
ext.abies = extraction(abies, climondbioclim, schema='flat', factor = 4);
dens.abies = densform(ext.abies, climondbioclim);
sea <- geo_findlocal(ext.abies, climondbioclim, 
 type = '.kde', maxiter = 5, 
 searchrep = 1, manip = 'condi', 
 divisions = 5)
}

