% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bison.R
\name{get_bison}
\alias{get_bison}
\title{Download distribution data from BISON (https://bison.usgs.gov)}
\usage{
get_bison(taxon, maxrec = Inf)
}
\arguments{
\item{taxon}{A string of the form 'genus species'.}

\item{maxrec}{Limit on number of records to download.}
}
\description{
This function requests data from the BISON database for a single species. 
Note that BISON requires exact name matching to binomial. Searching on a
genus, for example, will match only records that are enterred with that name
only, NOT all records in that genus identified by a binomial. Also note that
BISON has no name correction so mispelling and errors are likely.
}
\examples{
\dontrun{
abies <- get_bison('Abies fraseri', 10000);
}
}
