% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_direct.R
\name{get_dist_all}
\alias{get_dist_all}
\title{Download distribution data from BIEN, GBIF, Inaturalist,}
\usage{
get_dist_all(taxon, maxrec = 19999, local = FALSE, db = 0, h = 0,
  u = 0, pass = 0)
}
\arguments{
\item{taxon}{A string of the form 'genus species' or 'genus'.}

\item{maxrec}{Maximum number of records to download.}

\item{local}{TRUE or FALSE. To use the GBIF API use FALSE}

\item{db}{SQL database. ONLY FOR LOCAL GBIF INSALLATION}

\item{h}{SQL host}

\item{u}{SQL user}

\item{pass}{SQL password}
}
\description{
This function requests data from the GBIF database for a single taxon using the GBIF callback API.
}
\examples{
\dontrun{
abies <- gbif_dist_all('Abies', maxrec = 1000, local = FALSE);
}
}
