% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_data.R
\name{get_envirem_elev}
\alias{get_envirem_elev}
\title{Download ENVIREM topographical data set(s)}
\usage{
get_envirem_elev(period = "cur")
}
\arguments{
\item{period}{A string for time period to get. Either 'cur', 'midholo', or 'lgm'.}
}
\description{
This function requests raster layers from the envirem dataset (Bemmels, et al. 2017, Ecography)
downloads the zip archive, reads files into R and disposes of the originals because you should 
save the layers in an R ready raster object using raster::writeRaster().
}
\examples{
\dontrun{
#get 2.5 arcmin grid. 
envir <- get_envirem_elev(period='cur'); 
}
}
