% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cracle_build.R
\name{get_optim}
\alias{get_optim}
\title{Find PDF optim(a)um}
\usage{
get_optim(dens.ob)
}
\arguments{
\item{dens.ob}{An object derived from the cracle::dens_ob() function.}
}
\description{
Using an object from the cracle::dens_obj() function. Create a single density object (i.e., like that produced by cracle::densform()) where the probability curves correspond to the probability density function of ALL taxa/species from the original set occurring.
}
\examples{
#distr <- read.table('test_mat.txt', head=T, sep ="\\t");
#OR:
data(distr);
data(climondbioclim);
extr.raw = extraction(data=distr, clim= climondbioclim, schema='raw');
dens.list.raw <- dens_obj(extr.raw, clim = climondbioclim, bw = 'nrd0', n = 1024);
multiplot(dens.list.raw, names(climondbioclim[[1]]));
and <- and_fun(dens.list.raw);
addplot(and, names(climondbioclim[[1]]), col ='black');
optim.and <- get_optim(and);
abline(v=optim.and$means[paste(names(climondbioclim[[1]]), 'mean', sep = ".") ])
}

