% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_data.R
\name{get_worldclim}
\alias{get_worldclim}
\title{Download WorldClim data for modern or paleo models.}
\usage{
get_worldclim(period = "cur", model = "", version = "1_4",
  varset = "bio", res = 2.5)
}
\arguments{
\item{period}{A string. Either 'cur', 'midholo', or 'lgm'.}

\item{model}{For paleo models. Which to use (i.e., 'ccsm4'). See http://worldclim.org/paleo-climate1 for options.}

\item{version}{Either '1_4', or '2.0'}

\item{varset}{Either 'bio', 'tmean', 'tmin', 'tmax', or 'prec'.}

\item{res}{What spatial resolution? '10', '2.5' arcmin, or '30' seconds. (options are '10', '2.5', '30').}
}
\description{
This function requests climate raster objects from www.worldclim.org, 
downloads the files, reads them into R and disposes of the original files.
}
\examples{
\dontrun{
#get 10 minute mean monthly temperature grids.

abies <- get_worldclim(period='cur', varset = 'tmean', res=10); 
}
}
