% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rInat.R
\name{inat}
\alias{inat}
\title{Download distribution data from iNaturalist}
\usage{
inat(taxon, maxrec = 10000)
}
\arguments{
\item{taxon}{A string of the form 'genus species'.}

\item{maxrec}{Limit on number of records to download.}
}
\description{
This function requests data from the iNaturalist database for a single species that
wraps the data access function(s) from the rinat library.
}
\examples{
\dontrun{
abies <- inat('Abies fraseri', 10000);
}
}
