% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCR.R
\name{MCR}
\alias{MCR}
\title{A function to estimate climate using the Mutual Climatic Range method according to Thompson et al 2012.}
\usage{
MCR(ext, method = "unweight", plot = FALSE, file = "mcr_plot")
}
\arguments{
\item{ext}{An object generated by the extraction() function in this package or equivalent data.frame containing species IDs and climate data.}

\item{method}{A character string of value "weight" or "unweight" only. Defaults to "unweight" as this is the classic model.}

\item{plot}{Boolean (T or F) to indicate whether coexistence interval plots should be generated for each variable.}

\item{file}{If plot=T then provide a file name stub (prefix) to identify the plots to be generated. To this stub each variable name and the file extension will be appended.}
}
\description{
The Mutual Climatic Range method (see especially: Thompson et al 2012) and the Coexistence Approach are fundamentally similar methods implemented in this function to find the minimum interval of coexistence of a list of taxa. This function implements the MCR in both the weighted and unweighted form using a data.frame of climate data attached to species locality information.
}
\examples{
\dontrun{
data(distr);
data(climondbioclim);
extr.raw = extraction(data=distr, clim= climondbioclim, schema='raw');
mcr = MCR(extr.raw, method = 'unweight', plot=FALSE);
mcrw = MCR(extr.raw, method = 'weight', plot=TRUE, file = 'mcr_plot');
}
}
