% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cracle_build.R
\name{multiplot}
\alias{multiplot}
\title{Plot PDF curves of given type and variable for density list object}
\usage{
multiplot(dens.oblist, var, col = grDevices::heat.colors(length(dens.oblist)),
  type = ".kde", l.pos = "topleft", l.cex = 0.8, w = FALSE)
}
\arguments{
\item{dens.oblist}{An object derived from the vegdistmod::dens_obj() function.}

\item{var}{A character string that matches one of the layer names in the source raster object.}

\item{col}{A color vector of the same length as the dens.oblist object. Default is 'grDevices::heat.colors(length(dens.oblist))'.}

\item{type}{A character string of value either ".kde" for a Kernel Density Estimator curve, or ".gauss" for a Gaussian (normal) curve. All other values will result in errors.}

\item{l.pos}{Legend position. Recommend 'topleft' or 'topright'. Default is 'topleft'.}

\item{l.cex}{cex setting for legend. Default is 0.8.}

\item{w}{TRUE or FALSE to show weighted probability functions}
}
\description{
Using an object from the vegdistmod::dens_obj() function. Plot a series PDF curves in a new plot window.
}
\examples{
\dontrun{
#distr <- read.table('test_mat.txt', head=T, sep ="\\t");
#OR:
data(distr); data(climondbioclim);
extr.raw = extraction(data=distr, clim= climondbioclim, schema='raw');
dens.list.raw <- dens_obj(extr.raw, clim = climondbioclim, bw = 'nrd0', n = 1024);
multiplot(dens.list.raw, names(climondbioclim[[1]]));
}
}
