% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fun.R
\name{multiv_likelihood}
\alias{multiv_likelihood}
\title{The multivariate likelihood (log-likelihood) for a given set of PDF climate functions and localities.}
\usage{
multiv_likelihood(x, clim, dens, type, w = FALSE)
}
\arguments{
\item{x}{A data.frame of climate values (i.e., extracted from the climate raster object)}

\item{clim}{A raster object of climate data (matching x)}

\item{dens}{A density object from vegdistmod::densform(), vegdistmod::and_fun(), or vegdistmod::or_fun();}

\item{type}{Designate either ".gauss" or ".kde".}

\item{w}{To weight or not to weight by coefficient of variation.}
}
\description{
Returns a log-likelihood calculated from the climate data at one 
 or more localities referenced against a set of climate PDFs.
}
\examples{
data(abies);
ext.abies = extraction(abies, climondbioclim, schema='raw');
dens.abies = densform(ext.abies, climondbioclim);
for(i in 1:length(ext.abies[,1])){
  m = multiv_likelihood(ext.abies[i,6:length(ext.abies[1,])],
     climondbioclim, dens.abies, type = '.kde');
   ext.abies[i, 'prob'] = m;
}
print(ext.abies$prob)
}

