% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fun.R
\name{near2}
\alias{near2}
\title{Find nearby localities that are likely occurrences given environmental PDFs.}
\usage{
near2(ext_ob, clim, dens_ob, type, name = "NULL", w = FALSE)
}
\arguments{
\item{ext_ob}{A data.frame of climate values (i.e., extracted from the climate raster object). Will be passed to multiv_likelihood().}

\item{clim}{A raster object of climate data (matching x)}

\item{dens_ob}{A density object from vegdistmod::densform(), vegdistmod::and_fun(), or vegdistmod::or_fun();}

\item{type}{Designate either ".gauss" or ".kde".}

\item{name}{Optional. Give taxon name.}

\item{w}{To weight log-likelihoods by coefficient of variation or not.}
}
\description{
Given an occurrence set, find nearby localities that have an equal 
or greater multivariate likelihood given a set of environmental PDFs. 
Will return up to one simulated record per given occurrence, however,
 it is unlikely that so many suitable points will be found.
}
\examples{
data(abies);
ext.abies = extraction(abies, climondbioclim, schema='raw');
dens.abies = densform(ext.abies, climondbioclim);
n <- near2(ext.abies, climondbioclim, dens.abies, type = '.kde');
}

