% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNeotoma.R
\name{neotoma.sp}
\alias{neotoma.sp}
\title{Download distribution data from NEOTOMA}
\usage{
neotoma.sp(taxon)
}
\arguments{
\item{taxon}{A string of the form 'genus species' or 'genus'.}
}
\description{
This function requests data from the NEOTOMA database for a single taxon.
}
\examples{
\dontrun{
abies <- rNeotoma('Abies');
}
}
