% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fun.R
\name{plot_clim}
\alias{plot_clim}
\title{Plot lat/long points on a raster map.}
\usage{
plot_clim(ext_ob, clim, boundaries = "", file = "", col = "red")
}
\arguments{
\item{ext_ob}{A data.frame of climate values (i.e., extracted from the climate raster object). MUST include columns named 'lon' and 'lat'.}

\item{clim}{A raster object of climate data (matching ext_ob)}

\item{boundaries}{A shapefile (e.g., GADM country or state outlines)}

\item{file}{If a file path is set this function will try to write the plot as a png to that path}

\item{col}{Color of points to plot.}
}
\description{
Plotting with fancy colors. Canned so you don't have to think to hard about it.
}
\examples{
data(abies);
ext.abies = extraction(abies, climondbioclim, schema='raw');
plot_clim(ext.abies, climondbioclim[[5]]);

}

