% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{que.al}
\alias{que.al}
\title{Distribution of Quercus alba}
\format{A data.frame with 4 columns
'que.al'}
\usage{
data(que.al)
}
\description{
A sample of GBIF point locality occurrence data for Quercus alba for illustrating findlocal and geo_findlocal routines
}
\keyword{datasests}

