% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cracle_build.R
\name{write_results}
\alias{write_results}
\title{Write results to file}
\usage{
write_results(optima, siteval = "", clim, method, filename, append = F)
}
\arguments{
\item{optima}{A results object from vegdistmod::get_optim()}

\item{siteval}{A vector of site values for envirnomental parameters (See example for generating).}

\item{clim}{The original raster object used in the vegdistmod::extraction() step.}

\item{method}{The result summary method. Supported methods are "conintkde", "conintgauss", "origg", "origk", "dirconint" for the "get_optim()" type objects.}

\item{filename}{Path to desired file.}

\item{append}{Should this be appended to an existing file. Default is FALSE.}
}
\description{
Write results from a results object from either vegdistmod::get_optim() to a tab delimited table in a text file. This is done one 'method' at a time. Supported methods are "conintkde", "conintgauss", "origg", "origk", "dirconint" for the "get_optim()" type objects.
}
\examples{
\dontrun{
#distr <- read.table('test_mat.txt', head=T, sep ="\\t");
#OR:
data(distr); data(climondbioclim);
extr.raw = extraction(data=distr, clim= climondbioclim, schema='raw');
dens.list.raw <- dens_obj(extr.raw, clim = climondbioclim, bw = 'nrd0', n = 1024);
and <- and_fun(dens.list.raw);
optima <- get_optim(and);
#Compare to optima.
write_results(optima, clim = climondbioclim, method = 'conintkde', file = 'raw')
}
}

