% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cracle_build.R
\name{dens_obj}
\alias{dens_obj}
\title{A wrapper for vegdistmod::densform where a multi-taxon extraction object can be passed to densform one taxon at a time.}
\usage{
dens_obj(ex, clim, manip = "condi", bw = "nrd0", kern = "optcosine",
  clip = 0, n = 1024, parallel = FALSE, nclus = 4, bg.n = 200)
}
\arguments{
\item{ex}{An object derived from the extraction() function.}

\item{clim}{A raster object (see raster::raster() and raster::stack() documentation for reading raster files into R).}

\item{manip}{Character string of 'reg' for straight likelihood, 'condi' for conditional likelihood statement.}

\item{bw}{A bandwidth compatible with stats::density(). Options include "nrd", "nrd0", "ucv", "bcv", etc.. Default (and recommended) value is "nrd0".}

\item{kern}{Type of Kernel to smooth with. Recommend 'gaussian', 'optcosine', or 'epanechnikov'. See: stats::density for options.}

\item{clip}{A character string of value "range" or "95conf" or "99conf". Should the probability functions be clipped to either the empirical range or the 95 or 99 percent confidence interval?}

\item{n}{Number of equally spaced points at which the probability density is to be estimated. Defaults to 1024. A lower number increases speed but decreases resolution in the function. A higher number increases resolution at the cost of speed. Recommended values: 512, 1024, 2048, ....}

\item{parallel}{TRUE or FALSE. Make use of multicore architecture.}

\item{nclus}{Number of cores to allocate to this function}

\item{bg.n}{If there is not a background matrix, how many background points PER OCCURRENCE record should be sampled. Default is 1000.}
}
\description{
This function takes extracted climate data (from an object generated by the vegdistmod::extraction() function) and generates probability density functions for each taxon/variable pair using both a Gaussian (normal) approximation and a Gaussian Kernel Density estimator.
}
\examples{
\dontrun{
#distr <- read.table('test_mat.txt', head=T, sep ="\\t");
#OR:
data(distr);
data(climondbioclim);
extr.raw = extraction(data=distr, clim= climondbioclim, 
 schema='flat', factor = 4, rm.outlier=FALSE);
dens.list.raw <- dens_obj(extr.raw, clim = climondbioclim, 
 manip = 'condi', bg.n = 200, bw = 'nrd0', n = 1024);
multiplot(dens.list.raw, names(climondbioclim[[1]]));
}
}
