% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fun.R
\name{heat_up}
\alias{heat_up}
\title{Convert log-likelihoods of any vegdistmod PDF to a raster heat-map.}
\usage{
heat_up(clim, dens, parallel = FALSE, nclus = 4, type = ".kde",
  w = FALSE)
}
\arguments{
\item{clim}{A raster object of climate data (matching ext_ob)}

\item{dens}{A vegdistmod density object (see densform())}

\item{parallel}{Make use of multicore architecture}

\item{nclus}{If parallel is TRUE, how many cores should be allocated.}

\item{type}{Which PDF should be used, .gauss or .kde}

\item{w}{TRUE or FALSE should variable PDFs be weighted by relative niche breadth.}
}
\description{
This is not an ENM. What it does do is provides a geographic representation of the likelihood functions
 that are used in CRACLE and the spatial defragmentation functions.
}
\examples{
data(abies);
ext.abies = extraction(abies, climondbioclim, schema='raw');
plot_clim(ext.abies, climondbioclim[[5]]);

}

